Renaming Bot to Game
I realize that the concept I was looking for was "game" - the "bot" is
being handled by a separate part of the system, and my code isn't the
bot.  My code is the game that is being called by the bot.

Once I had the right name for the concept, I could start correctly
assigning methods and attributes to the object.  A game is where we will
be keeping track of which pools a player has (at least until we
implement the concept of a `WickedConflict`, which is the next major
refactoring.)