Dice pools shouldn't care about the player
A dice pool shouldn't be keeping track of the player who rolled it.  It
should be maintaining the list of dice in the pool, and the result of
those dice.  But who rolled the dice should (for now) be handled by
whatever object called for building the dice pool.  (as of now, this is
the `WickedGame` object, but in the future it may be `WickedConflict`.)