Add metaprogramming for list of commands
The process of adding commands was going to start getting tedious.
Having a separate block for each command defining what to call in
`WickedGame` for this command was going to be either very repetitive
since each block would be identical, or it would be introducing code
that would be hard to have under test.

Instead, this model has `WickedGame` share a list of commands that it
can handle, and `wicked.rb` just passes the event invocation details to
`WickedGame` for all commands in that list.  With this change, we can
add new methods to `WickedGame` and just add them into `COMMANDS`, and
they will be automatically stitched into the list of commands that
Discordrb is listening for.