Implementing `roll_for`
Allows someone to roll for a player or NPC by giving a name followed by
a dice array.  Note, if the dice array includes an advantage die, that
die cannot be at the start of the dice array list.  All words up until
the first die will be considered part of the name.  Anything after the
first die will be considered part of the dice pool.