Add advantage dice modification
It is now possible to use `/advantage` to adjust advantage dice in a
pool.  Specifically, you can now say `/advantage +` to add and advantage
die, or `/advantage -` to remove one from your pool.  You can also use
`/advantage Some Character Name +` to add an advantage die to Some
Character Name's pool, and similar for `-`.

Note that `/advantage` is using a pattern I hope to extend to the other
`/` commands which have regular versions and `_for` versions where you
specify the character name.  I want to change them so that they all will
check to see if there is a character name.  But that is a future
enhancement.