Fix a list bug with unrolled advantage dice
When a player had added an advantage die, but had not rolled it yet, the
logic for displaying the list of dice would break as it tried to add the
advantage dice together, but the dice had the results of "unrolled".

The fix was to break apart the concept of "result" and "value".  The
result of a die is "unrolled" - the string representation, basically.
The value of a die is 0 if it hasn't been rolled.

By separating this out, we can identify when we want the values (when we
are summing up the total result) and when we want the result (when we
are displaying the current pool state.

Most of the work here was adding the tests, which are possibly
excessive, but it was helpful to step down through the code to identify
where the failure was happening, so I'm keeping all of the tests for
now.