Remove `_for` methods
It was hard to remember to type _for all of the time when working as the
GM.  Instead, it makes sense to just have the system figure out if
you've passed a player name.  If there is a player name, then use that
name.  If there isn't, then use the default name passed by the event.

This cleaned up the code a good bit, really.  I like this refactoring -
it removes a lot of duplicate code.