Fix feature test so it passes correctly
This feature has been working for a while.  However, the pending tag
kept this failing test from being noticed.

The problem was that using an injected randomizer doesn't work after you
have created the pool.  The randomizer is used when instantiating the
die pool, but after that it isn't used for new rolls.  This means the
injected randomizer stubbing wasn't have any effect, so the results we
were getting didn't match the expectations.

However, there is no reason to worry about what values we get back on
this test.  The test doesn't care what the values of the dice are, so
long as there are some values.  Rewriting the tests to reflect that
allow us to remove the stub object.