Sorted player pools
This was easier than I expected, because it turns out the Array already
is comparable, and it implements the <=> operator exactly as I wanted.

Because of that, all we need to do is pass the comparable request to the
result, which has the actual values, and let it handle the comparison
with the value arrays.  To ensure that they are compared correctly, we
need to sort them before the comparison, so it makes sense to pull that
out since we now have two places we are using the sorted values.