Fix crashes caused by bad input
The system was interpreting most bad input as if it was a character,
with no dice.  It was then throwing an error because, if there is a
character name provided without any dice, it assumes the character
already has a dice pool that should be re-rolled.

However, in this case, there is no dice pool.  The solution is to add a
guard clause that will ensure that the input is what we are expecting,
and to give an error if it doesn't match.