Partial name lookup
It was hard to always type out the long name of a character exactly.
Now, after giving the name for the first roll, a player can just type in
the start of the name and if there is a character who already matches
that name in the system, it will use that name.  If there isn't a
matching name, then it will treat the name as a new character.  If there
are two matching names already in the system, it will return an error
message, indicating that there is a duplicate match.